# backend/src/main/java/com/kombaos/inventory/threshold/domain/MaterialStockThreshold.java
package com.kombaos.inventory.threshold.domain;

import java.math.BigDecimal;
import java.time.Instant;

public record MaterialStockThreshold(
        String materialId,
        BigDecimal minStock,
        Instant updatedAt
) {
}