# shared/openapi/kombaos.yaml
openapi: 3.0.3
info:
  title: KombaOS API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/materials:
    get:
      summary: List materials
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: supplier
          required: false
          schema:
            type: string
        - in: query
          name: origin
          required: false
          schema:
            type: string
        - in: query
          name: certified
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Material'
    post:
      summary: Create material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpsert'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
  /api/materials/{id}:
    get:
      summary: Get material
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '404':
          description: Not found
    put:
      summary: Update material
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpsert'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
    delete:
      summary: Delete material
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /api/materials/{materialId}/stock:
    get:
      summary: Get material stock
      parameters:
        - in: path
          name: materialId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialStock'
        '404':
          description: Not found
  /api/materials/{materialId}/threshold:
    get:
      summary: Get material stock threshold
      parameters:
        - in: path
          name: materialId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialStockThreshold'
        '404':
          description: Not found
    put:
      summary: Upsert material stock threshold
      parameters:
        - in: path
          name: materialId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialStockThresholdUpsert'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialStockThreshold'
        '400':
          description: Bad request
        '404':
          description: Not found
    delete:
      summary: Delete material stock threshold
      parameters:
        - in: path
          name: materialId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /api/inventory/movements:
    get:
      summary: List inventory movements
      parameters:
        - in: query
          name: materialId
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryMovement'
    post:
      summary: Create inventory movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryMovementCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryMovement'
        '400':
          description: Bad request
        '404':
          description: Not found
  /api/inventory/movements/{id}:
    get:
      summary: Get inventory movement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryMovement'
        '404':
          description: Not found
    delete:
      summary: Delete inventory movement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /api/inventory/alerts/low-stock:
    get:
      summary: List low stock alerts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LowStockAlert'
components:
  schemas:
    Material:
      type: object
      required: [id, name, unit, certified, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        unit:
          type: string
        supplier:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        certified:
          type: boolean
        costCents:
          type: integer
          format: int64
          nullable: true
        currency:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    MaterialUpsert:
      type: object
      required: [name, unit]
      properties:
        name:
          type: string
        unit:
          type: string
        supplier:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        certified:
          type: boolean
          nullable: true
        costCents:
          type: integer
          format: int64
          nullable: true
        currency:
          type: string
          nullable: true
    MaterialStock:
      type: object
      required: [materialId, stock]
      properties:
        materialId:
          type: string
        stock:
          type: number
    MaterialStockThreshold:
      type: object
      required: [materialId, minStock, updatedAt]
      properties:
        materialId:
          type: string
        minStock:
          type: number
        updatedAt:
          type: string
          format: date-time
    MaterialStockThresholdUpsert:
      type: object
      required: [minStock]
      properties:
        minStock:
          type: number
    LowStockAlert:
      type: object
      required: [materialId, name, unit, stock, minStock]
      properties:
        materialId:
          type: string
        name:
          type: string
        unit:
          type: string
        stock:
          type: number
        minStock:
          type: number
    InventoryMovementType:
      type: string
      enum: [IN, OUT, ADJUST]
    InventoryMovement:
      type: object
      required: [id, materialId, type, quantity, createdAt]
      properties:
        id:
          type: string
        materialId:
          type: string
        type:
          $ref: '#/components/schemas/InventoryMovementType'
        quantity:
          type: number
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
    InventoryMovementCreate:
      type: object
      required: [materialId, type, quantity]
      properties:
        materialId:
          type: string
        type:
          $ref: '#/components/schemas/InventoryMovementType'
        quantity:
          type: number
        reason:
          type: string